Cache layer added but there are a lot of bugs and issues

+++ MyExams.Models/Test.cs
@@ -20,7 +20,7 @@ namespace MyExams.Models
         public string TestTitle { get; set; }
         public DateTime LastUpdated { get; set; }
         public int Students { get; set; }
-        public double AverageMark { get; set; }
+        public int AverageMark { get; set; }
         public DateTime RecentUsage { get; set; }
     }
 }
